requirements.txt
langchain
langchain-google-genai
duckduckgo-search
langchain-community
streamlit